+++ lib/Latte/CoreFiltersExtension.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace JanHerman\Barista\Latte;
+
+use Latte\Extension;
+
+class CoreFiltersExtension extends Extension
+{
+    public function getFilters(): array
+    {
+        return [
+            'stripNewLines' => Filters::stripNewLines(...),
+        ];
+    }
+}